rebase modal for ingredient
