+++ src/components/burger-ingredients-item/burger-ingredients-item.jsx
@@ -1,15 +1,11 @@
-import { useMemo, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useMemo } from "react";
+import { useSelector } from "react-redux";
 import { useDrag } from "react-dnd";
-import Modal from "../modal";
-import IngredientDetails from "../ingredient-details";
 import { CurrencyIcon, Counter } from "@ya.praktikum/react-developer-burger-ui-components";
 import PropTypes from "prop-types";
 import style from "./burger-ingredients-item.module.css";
-import { setDetailInfo, deleteDetailInfo } from "../../services/action-creators/ingredient";
 
-const BurgerIngredientsItem = ({ data }) => {
-  const dispatch = useDispatch();
+const BurgerIngredientsItem = ({ data, openModal }) => {
   const { constructorIngredient, bun } = useSelector((state) => state.constructorIngredient);
 
   const count = useMemo(() => {
@@ -21,41 +17,21 @@ const BurgerIngredientsItem = ({ data }) => {
     return 0;
   }, [bun, constructorIngredient, data._id, data.type]);
 
-  const [showModal, setShowModal] = useState(false);
   const [, ref] = useDrag({
     type: "ingredient",
     item: data,
   });
 
   return (
-    <>
-      <div
-        className={style.wrapper}
-        onClick={() => {
-          dispatch(setDetailInfo({ ingredientDetail: data }));
-          setShowModal(true);
-        }}
-      >
-        {!!count && <Counter count={count} size="default" />}
-        <img ref={ref} className={style.image} src={data.image} />
-        <div className={style.price_wrapper}>
-          <p className="text text_type_digits-default mr-2">{data.price}</p>
-          <CurrencyIcon type="primary" />
-        </div>
-        <p className={`${style.description} text text_type_main-default`}>{data.name}</p>
+    <div className={style.wrapper} onClick={openModal}>
+      {!!count && <Counter count={count} size="default" />}
+      <img ref={ref} className={style.image} src={data.image} />
+      <div className={style.price_wrapper}>
+        <p className="text text_type_digits-default mr-2">{data.price}</p>
+        <CurrencyIcon type="primary" />
       </div>
-      {showModal && (
-        <Modal
-          header="Детали ингредиента"
-          onDestroyModal={() => {
-            setShowModal(false);
-            dispatch(deleteDetailInfo());
-          }}
-        >
-          <IngredientDetails />
-        </Modal>
-      )}
-    </>
+      <p className={`${style.description} text text_type_main-default`}>{data.name}</p>
+    </div>
   );
 };
 
@@ -67,6 +43,7 @@ BurgerIngredientsItem.propTypes = {
     type: PropTypes.string,
     _id: PropTypes.string,
   }),
+  openModal: PropTypes.func,
 };
 
 export default BurgerIngredientsItem;
